+++ src/types/SignatureRate.ts
@@ -1,0 +1,4 @@
+export interface SignatureRate {
+  _id: string;
+  participationRate: number;
+}