+++ CHANGELOG.md
@@ -5,6 +5,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+- Added `GET: /signatures` to get participation rate of each voter in given interval of blocks
 
 ## [4.18.0] - 2022-05-05
 ### Added